FEM: import tools, displacement abs, move calculation and adding into result tools
