+++ src/Mod/Fem/feminout/importVTKResults.py
@@ -119,7 +119,6 @@ def importVtkFCResult(filename, resultname, analysis=None, result_name_prefix=No
     # See _getFreeCADMechResultProperties() in FemVTKTools.cpp for the supported names
 
     import ObjectsFem
-    from . import importToolsFem
     if result_name_prefix is None:
         result_name_prefix = ''
     if analysis:
@@ -129,10 +128,10 @@ def importVtkFCResult(filename, resultname, analysis=None, result_name_prefix=No
     result_obj = ObjectsFem.makeResultMechanical(FreeCAD.ActiveDocument, results_name)
     Fem.readResult(filename, result_obj.Name)  # readResult always creates a new femmesh named ResultMesh
 
-    # workaround for the DisplacementLengths (They should have been calculated by Fem.readResult)
+    # add missing DisplacementLengths (They should have been added by Fem.readResult)
     if not result_obj.DisplacementLengths:
-        result_obj.DisplacementLengths = importToolsFem.calculate_disp_abs(result_obj.DisplacementVectors)
-        FreeCAD.Console.PrintMessage('Recalculated DisplacementLengths.\n')
+        import femresult.resulttools as restools
+        result_obj = restools.add_disp_apps(result_obj)  # DisplacementLengths
 
     ''' seems unused at the moment
     filenamebase = '.'.join(filename.split('.')[:-1])  # pattern: filebase_timestamp.vtk
